move logic in dashboard selectors to meta/Dashboard
