+++ frontend/src/metabase/dashboard/selectors.js
@@ -1,13 +1,14 @@
 import _ from "underscore";
-import { setIn } from "icepick";
 
 import { createSelector } from "reselect";
 
 import { getMetadata } from "metabase/selectors/metadata";
 
-import * as Dashboard from "metabase/meta/Dashboard";
+import {
+  getParameterMappingOptions as _getParameterMappingOptions,
+  getMappingsByParameter as _getMappingsByParameter,
+} from "metabase/meta/Dashboard";
 
-import { getParameterTargetFieldId } from "metabase/meta/Parameter";
 import { SIDEBAR_NAME } from "metabase/dashboard/constants";
 
 import type { CardId, Card } from "metabase-types/types/Card";
@@ -138,87 +139,7 @@ export const getParameterTarget = createSelector(
 
 export const getMappingsByParameter = createSelector(
   [getMetadata, getDashboardComplete],
-  (metadata, dashboard) => {
-    if (!dashboard) {
-      return {};
-    }
-
-    let mappingsByParameter: MappingsByParameter = {};
-    const mappings: AugmentedParameterMapping[] = [];
-    const countsByParameter = {};
-    for (const dashcard of dashboard.ordered_cards) {
-      const cards: Card[] = [dashcard.card].concat(dashcard.series);
-      for (const mapping: ParameterMapping of dashcard.parameter_mappings ||
-        []) {
-        const card = _.findWhere(cards, { id: mapping.card_id });
-        const fieldId =
-          card && getParameterTargetFieldId(mapping.target, card.dataset_query);
-        const field = metadata.field(fieldId);
-        const values = (field && field.fieldValues()) || [];
-        if (values.length) {
-          countsByParameter[mapping.parameter_id] =
-            countsByParameter[mapping.parameter_id] || {};
-        }
-        for (const value of values) {
-          countsByParameter[mapping.parameter_id][value] =
-            (countsByParameter[mapping.parameter_id][value] || 0) + 1;
-        }
-
-        const augmentedMapping: AugmentedParameterMapping = {
-          ...mapping,
-          parameter_id: mapping.parameter_id,
-          dashcard_id: dashcard.id,
-          card_id: mapping.card_id,
-          field_id: fieldId,
-          values,
-        };
-        mappingsByParameter = setIn(
-          mappingsByParameter,
-          [mapping.parameter_id, dashcard.id, mapping.card_id],
-          augmentedMapping,
-        );
-        mappings.push(augmentedMapping);
-      }
-    }
-    const mappingsWithValuesByParameter = {};
-    // update max values overlap for each mapping
-    for (const mapping of mappings) {
-      if (mapping.values && mapping.values.length > 0) {
-        const overlapMax = Math.max(
-          ...mapping.values.map(
-            value => countsByParameter[mapping.parameter_id][value],
-          ),
-        );
-        mappingsByParameter = setIn(
-          mappingsByParameter,
-          [
-            mapping.parameter_id,
-            mapping.dashcard_id,
-            mapping.card_id,
-            "overlapMax",
-          ],
-          overlapMax,
-        );
-        mappingsWithValuesByParameter[mapping.parameter_id] =
-          (mappingsWithValuesByParameter[mapping.parameter_id] || 0) + 1;
-      }
-    }
-    // update count of mappings with values
-    for (const mapping of mappings) {
-      mappingsByParameter = setIn(
-        mappingsByParameter,
-        [
-          mapping.parameter_id,
-          mapping.dashcard_id,
-          mapping.card_id,
-          "mappingsWithValues",
-        ],
-        mappingsWithValuesByParameter[mapping.parameter_id] || 0,
-      );
-    }
-
-    return mappingsByParameter;
-  },
+  _getMappingsByParameter,
 );
 
 /** Returns the dashboard's parameters objects, with field_id added, if appropriate */
@@ -267,7 +188,7 @@ export const makeGetParameterMappingOptions = () => {
       parameter: Parameter,
       card: Card,
     ): Array<ParameterMappingUIOption> => {
-      return Dashboard.getParameterMappingOptions(metadata, parameter, card);
+      return _getParameterMappingOptions(metadata, parameter, card);
     },
   );
   return getParameterMappingOptions;
